DS: Move graphics code into a separate file
