+++ backends/platform/ds/dsmain.h
@@ -23,51 +23,14 @@
 #ifndef _DSMAIN_H
 #define _DSMAIN_H
 
-#define FORBIDDEN_SYMBOL_ALLOW_ALL
-
-#include <nds.h>
-#include "osystem_ds.h"
+#include "common/scummsys.h"
 
 namespace DS {
-
-// Video
-void 	displayMode8Bit();											// Switch to 8-bit mode5
-void 	displayMode16Bit();										// Switch to 16-bit mode5
-
-// Get address of current back buffer
-u16 *	get16BitBackBuffer();
-
-void 	setTalkPos(int x, int y);
-void 	setTopScreenTarget(int x, int y);
-void	setTopScreenZoom(int percentage);
-
-// Events
-void 	VBlankHandler();
-Common::Point transformPoint(uint16 x, uint16 y, bool isOverlayShown);
-Common::Point warpMouse(int penX, int penY, bool isOverlayShown);
-
-// Shake
-void 	setShakePos(int shakeXOffset, int shakeYOffset);
-
-// Options
-void	setGameScreenSwap(bool enable);
-
-// Display
-bool 	getIsDisplayMode8Bit();
-void 	setGameSize(int width, int height);
-int		getGameWidth();
-int		getGameHeight();
-void 	initHardware();
 
 // Fast RAM allocation (ITCM)
 void	fastRamReset();
 void*	fastRamAlloc(int size);
 
-
 } // End of namespace DS
-
-
-
-int cygprofile_getHBlanks();
 
 #endif
